+++ Sources/CLI/MainCLI.swift
@@ -19,7 +19,7 @@ import ArgumentParser
                     ScanCommand.self,
                     ListCommand.self,
                     CloneCommand.self,
-                    TransferCommand.self,
+                    RegisterCommand.self,
                     OpenCommand.self,
                     SelfUpdateCommand.self,
                 ],